+++ src/Resources/contao/languages/de/tl_chronik.php
@@ -3,80 +3,48 @@
 /**
  * Backend-Modul: Übersetzungen im Eingabeformular
  */
-$GLOBALS["TL_LANG"]["tl_chronik"]["chronik_legend"] = "Ereignis";
+$GLOBALS['TL_LANG']['tl_chronik']['chronik_legend'] = 'Ereignis';
+$GLOBALS['TL_LANG']['tl_chronik']['region'] = array('Ort', 'Geben Sie hier den Ort des Ereignisses ein.');
+$GLOBALS['TL_LANG']['tl_chronik']['from_date'] = array('Von', 'Von-Datum im Format JJJJ, MM.JJJJ oder TT.MM.JJJJ');
+$GLOBALS['TL_LANG']['tl_chronik']['to_date'] = array('Bis', 'Bis-Datum im Format JJJJ, MM.JJJJ oder TT.MM.JJJJ. Leerlassen bei einem eintägigen Ereignis.');
+$GLOBALS['TL_LANG']['tl_chronik']['title'] = array('Titel', 'Geben Sie hier den Titel des Ereignisses ein.');
+$GLOBALS['TL_LANG']['tl_chronik']['text'] = array('Teaser', 'Geben Sie hier einen kurzen Einführungstext ein.');
 
-$GLOBALS["TL_LANG"]["tl_chronik"]["region"][0] = "Ort";
-$GLOBALS["TL_LANG"]["tl_chronik"]["region"][1] = "Geben Sie hier den Ort des Ereignisses ein.";
+$GLOBALS['TL_LANG']['tl_chronik']['source_legend'] = 'Quelle';
+$GLOBALS['TL_LANG']['tl_chronik']['source'] = array('Quelle', 'Geben Sie hier den Namen der Quelle ein.');
+$GLOBALS['TL_LANG']['tl_chronik']['url'] = array('URL', 'Geben Sie hier einen Link zur Quelle ein.');
 
-$GLOBALS["TL_LANG"]["tl_chronik"]["from_date"][0] = "Von";
-$GLOBALS["TL_LANG"]["tl_chronik"]["from_date"][1] = "Von-Datum im Format JJJJ, MM.JJJJ oder TT.MM.JJJJ";
+$GLOBALS['TL_LANG']['tl_chronik']['image_legend'] = 'Bild';
+$GLOBALS['TL_LANG']['tl_chronik']['addImage'] = array('Bild hinzufügen', 'Fügen Sie hier ein Bild zum Teaser hinzu.');
+$GLOBALS['TL_LANG']['tl_chronik']['singleSRC'] = array('Quelldatei', 'Bitte wählen Sie einBild/Logo aus der Dateiübersicht.');
+$GLOBALS['TL_LANG']['tl_chronik']['size'] = array('Bildbreite und Bildhöhe', 'Hier können Sie die Abmessungen des Bildes und den Skalierungsmodus festlegen.');
+$GLOBALS['TL_LANG']['tl_chronik']['floating'] = array('Bildausrichtung', 'Bitte legen Sie fest, wie das Bild ausgerichtet werden soll.');
+$GLOBALS['TL_LANG']['tl_chronik']['fullsize'] = array('Großansicht/Neues Fenster', 'Großansicht des Bildes in einer Lightbox bzw. den Link in einem neuem Browserfenster öffnen.');
+$GLOBALS['TL_LANG']['tl_chronik']['overwriteMeta'] = array('Metadaten überschreiben', 'Metadaten überschreiben.');
+$GLOBALS['TL_LANG']['tl_chronik']['alt'] = array('Alternativer Text', 'Hier können Sie einen alternativen Text für das Bild eingeben (<i>alt</i>-Attribut)');
+$GLOBALS['TL_LANG']['tl_chronik']['imageUrl'] = array('Bildlink-Adresse', 'Eine eigene Bildlink-Adresse überschreibt den Lightbox-Link, so daß das Bild nicht mehr in der Großansicht dargestellt werden kann.');
+$GLOBALS['TL_LANG']['tl_chronik']['imageTitle'] = array('Bildtitel', 'Titel-Tag für das Bild');
+$GLOBALS['TL_LANG']['tl_chronik']['caption'] = array('Bildunterschrift', 'Hier können Sie einen kurzen Text eingeben, der unterhalb des Bildes angezeigt wird.');
 
-$GLOBALS["TL_LANG"]["tl_chronik"]["to_date"][0] = "Bis";
-$GLOBALS["TL_LANG"]["tl_chronik"]["to_date"][1] = "Bis-Datum im Format JJJJ, MM.JJJJ oder TT.MM.JJJJ. Leerlassen bei einem eintägigen Ereignis.";
+$GLOBALS['TL_LANG']['tl_chronik']['connect_legend'] = 'Verknüpfungen';
+$GLOBALS['TL_LANG']['tl_chronik']['clublist_incomplete'] = array('Vereine unzugeordnet', 'Vereine sind diesem Ereignis noch nicht komplett zugeordnet.');
+$GLOBALS['TL_LANG']['tl_chronik']['clublist_failed'] = array('Unzugeordnete Vereine', 'Diese Vereine konnten diesem Ereignis noch nicht zugeordnet werden.');
+$GLOBALS['TL_LANG']['tl_chronik']['clublist'] = array('Zugeordnete Vereine', 'Vereine diesem Ereignis zuordnen.');
+$GLOBALS['TL_LANG']['tl_chronik']['playerlist_incomplete'] = array('Spieler unzugeordnet', 'Spieler sind diesem Ereignis noch nicht komplett zugeordnet.');
+$GLOBALS['TL_LANG']['tl_chronik']['playerlist_failed'] = array('Unzugeordnete Spieler', 'Diese Spieler konnten diesem Ereignis noch nicht zugeordnet werden.');
+$GLOBALS['TL_LANG']['tl_chronik']['playerlist'] = array('Zugeordnete Spieler', 'Spieler diesem Ereignis zuordnen.');
 
-$GLOBALS["TL_LANG"]["tl_chronik"]["title"][0] = "Titel";
-$GLOBALS["TL_LANG"]["tl_chronik"]["title"][1] = "Geben Sie hier den Titel des Ereignisses ein.";
-
-$GLOBALS["TL_LANG"]["tl_chronik"]["text"][0] = "Teaser";
-$GLOBALS["TL_LANG"]["tl_chronik"]["text"][1] = "Geben Sie hier einen kurzen Einführungstext ein.";
-
-$GLOBALS["TL_LANG"]["tl_chronik"]["source_legend"] = "Quelle";
-
-$GLOBALS["TL_LANG"]["tl_chronik"]["source"][0] = "Quelle";
-$GLOBALS["TL_LANG"]["tl_chronik"]["source"][1] = "Geben Sie hier den Namen der Quelle ein.";
-
-$GLOBALS["TL_LANG"]["tl_chronik"]["url"][0] = "URL";
-$GLOBALS["TL_LANG"]["tl_chronik"]["url"][1] = "Geben Sie hier einen Link zur Quelle ein.";
-
-$GLOBALS["TL_LANG"]["tl_chronik"]["image_legend"] = "Bild";
-
-$GLOBALS["TL_LANG"]["tl_chronik"]["addImage"][0] = "Bild hinzufügen";
-$GLOBALS["TL_LANG"]["tl_chronik"]["addImage"][1] = "Fügen Sie hier ein Bild zum Teaser hinzu.";
-
-$GLOBALS["TL_LANG"]["tl_chronik"]["connect_legend"] = "Verknüpfungen";
-
-$GLOBALS["TL_LANG"]["tl_chronik"]["clublist_incomplete"][0] = "Vereine unzugeordnet";
-$GLOBALS["TL_LANG"]["tl_chronik"]["clublist_incomplete"][1] = "Vereine sind diesem Ereignis noch nicht komplett zugeordnet.";
-
-$GLOBALS["TL_LANG"]["tl_chronik"]["clublist_failed"][0] = "Unzugeordnete Vereine";
-$GLOBALS["TL_LANG"]["tl_chronik"]["clublist_failed"][1] = "Diese Vereine konnten diesem Ereignis noch nicht zugeordnet werden.";
-
-$GLOBALS["TL_LANG"]["tl_chronik"]["clublist"][0] = "Zugeordnete Vereine";
-$GLOBALS["TL_LANG"]["tl_chronik"]["clublist"][1] = "Vereine diesem Ereignis zuordnen.";
-
-$GLOBALS["TL_LANG"]["tl_chronik"]["playerlist_incomplete"][0] = "Spieler unzugeordnet";
-$GLOBALS["TL_LANG"]["tl_chronik"]["playerlist_incomplete"][1] = "Spieler sind diesem Ereignis noch nicht komplett zugeordnet.";
-
-$GLOBALS["TL_LANG"]["tl_chronik"]["playerlist_failed"][0] = "Unzugeordnete Spieler";
-$GLOBALS["TL_LANG"]["tl_chronik"]["playerlist_failed"][1] = "Diese Spieler konnten diesem Ereignis noch nicht zugeordnet werden.";
-
-$GLOBALS["TL_LANG"]["tl_chronik"]["playerlist"][0] = "Zugeordnete Spieler";
-$GLOBALS["TL_LANG"]["tl_chronik"]["playerlist"][1] = "Spieler diesem Ereignis zuordnen.";
-
-$GLOBALS["TL_LANG"]["tl_chronik"]["published_legend"] = "Veröffentlichung";
-
-$GLOBALS["TL_LANG"]["tl_chronik"]["published"][0] = "Veröffentlicht";
-$GLOBALS["TL_LANG"]["tl_chronik"]["published"][1] = "Ereignis aktivieren oder deaktivieren";
+$GLOBALS['TL_LANG']['tl_chronik']['publish_legend'] = 'Veröffentlichung';
+$GLOBALS['TL_LANG']['tl_chronik']['published'] = array('Veröffentlicht', 'Ereignis aktivieren oder deaktivieren');
 
 /**
  * Buttons für Operationen
  */
 
-$GLOBALS['TL_LANG']['tl_chronik']['new'][0] = "Neues Ereignis";
-$GLOBALS['TL_LANG']['tl_chronik']['new'][1] = "Neues Ereignis anlegen";
-
-$GLOBALS['TL_LANG']['tl_chronik']['edit'][0] = "Ereignis bearbeiten";
-$GLOBALS['TL_LANG']['tl_chronik']['edit'][1] = "Ereignis %s bearbeiten";
-
-$GLOBALS['TL_LANG']['tl_chronik']['copy'][0] = "Ereignis kopieren";
-$GLOBALS['TL_LANG']['tl_chronik']['copy'][1] = "Ereignis %s kopieren";
-
-$GLOBALS['TL_LANG']['tl_chronik']['delete'][0] = "Ereignis löschen";
-$GLOBALS['TL_LANG']['tl_chronik']['delete'][1] = "Ereignis %s löschen";
-
-$GLOBALS['TL_LANG']['tl_chronik']['toggle'][0] = "Ereignis aktivieren/deaktivieren";
-$GLOBALS['TL_LANG']['tl_chronik']['toggle'][1] = "Ereignis %s aktivieren/deaktivieren";
-
-$GLOBALS['TL_LANG']['tl_chronik']['show'][0] = "Ereignisdetails anzeigen";
-$GLOBALS['TL_LANG']['tl_chronik']['show'][1] = "Details des Ereignisses %s anzeigen";
+$GLOBALS['TL_LANG']['tl_chronik']['new'] = array('Neues Ereignis', 'Neues Ereignis anlegen');
+$GLOBALS['TL_LANG']['tl_chronik']['edit'] = array('Ereignis bearbeiten', 'Ereignis %s bearbeiten');
+$GLOBALS['TL_LANG']['tl_chronik']['copy'] = array('Ereignis kopieren', 'Ereignis %s kopieren');
+$GLOBALS['TL_LANG']['tl_chronik']['delete'] = array('Ereignis löschen', 'Ereignis %s löschen');
+$GLOBALS['TL_LANG']['tl_chronik']['toggle'] = array('Ereignis aktivieren/deaktivieren', 'Ereignis %s aktivieren/deaktivieren');
+$GLOBALS['TL_LANG']['tl_chronik']['show'] = array('Ereignisdetails anzeigen', 'Details des Ereignisses %s anzeigen');