+++ src/Resources/contao/languages/de/tl_module.php
@@ -1,0 +1,20 @@
+<?php
+
+/**
+ * Contao Open Source CMS
+ *
+ * Copyright (c) 2005-2014 Leo Feyer
+ *
+ * @package   fh-counter
+ * @author    Frank Hoppe
+ * @license   GNU/LGPL
+ * @copyright Frank Hoppe 2014
+ */
+
+$GLOBALS['TL_LANG']['tl_module']['lizenzverwaltung_legend'] = 'Einstellungen';
+
+$GLOBALS['TL_LANG']['tl_module']['chronik_from'] = array('Von Jahr', 'Jahr im Format JJJJ, bei dem die Chronik beginnen soll.');
+$GLOBALS['TL_LANG']['tl_module']['chronik_to'] = array('Bis Jahr', 'Jahr im Format JJJJ, bei dem die Chronik enden soll.');
+
+$GLOBALS['TL_LANG']['tl_module']['chronik_filter'] = array('Jahres-Menü statt Chronik anzeigen', 'Nur ein Jahres-Menü anzeigen. Die Chroniken werden auf virtuellen Unterseiten angezeigt.');
+$GLOBALS['TL_LANG']['tl_module']['chronik_timerange'] = array('Anzahl Jahre', 'Anzahl der Jahre je Menüpunkt.');