+++ src/Resources/contao/languages/de/tl_settings.php
@@ -1,0 +1,8 @@
+<?php
+
+/**
+ * Backend-Modul: Übersetzungen im Eingabeformular
+ */
+$GLOBALS['TL_LANG']['tl_settings']['chronik_legend'] = 'Chronik';
+$GLOBALS['TL_LANG']['tl_settings']['chronik_seite'] = array('Chronik-Modul-Seite', 'Geben Sie hier die Seite an, wo das Chronik-Modul eingebunden ist.');
+