+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Schachchronik Changelog
 
+## Version 1.1.0 (2023-11-30)
+
+* Add: Übersetzungen im Backend
+* Add: Einstellungen FE-Modul Chronik -> Ausgabe des Jahresmenüs jetzt optional mit Möglichkeit der Begrenzung des Zeitraums.
+* Change: FE-Modul Chronik von Classes nach Modules verschoben
+* Add: Abhängigkeit codefog/contao-haste
+* Change: Toggle-Funktion durch Haste-Toggler ersetzt
+* Add: Freigabe für PHP 8
+* Change: Bildeinbindung erneuert mit overwriteMeta
+* Change: tl_chronik.getSpieler ersetzt durch Funktionsaufruf Spielerregister
+* Change: Template chronik_list -> mod_chronik
+* Add: tl_settings mit der Seite, auf der das Chronikmodul eingebunden ist
+* Change: ALIAS_CHRONIK ersetzt durch tl_settings
+
 ## Version 1.0.2 (2019-09-28)
 
 * Fix: tl_chronik Helper-Links